upload-ready/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pymupdf>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
